+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 * `ProgIter` now correctly checks if it needs to displays a message on every iteration.
+* fixed uninitialized `_cursor_at_newline ` variable in `ProgIter`.
 
 ### Added
 * `ub.repr2` now accept type name strings at register time (which makes it